ELEC-376: split configs for front and back to their own files, added comments for more documentation.
